requirements.txt
google-cloud-storage
kfp
gspread
pyyaml
